hooks/hook-numpy.py
# 收集 numpy 的所有子模块和数据文件
from PyInstaller.utils.hooks import collect_submodules, collect_data_files

hiddenimports = collect_submodules('numpy')
datas = collect_data_files('numpy')

