requirements.txt
# CAT1设备测试平台依赖包
PyQt5==5.15.9
pyserial==3.5
pandas==2.0.3
numpy==1.24.3
xlsxwriter==3.1.2
matplotlib==3.7.2
psutil==5.9.5
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
pyusb>=1.2.0